refactoring
- orginized all in an MVC pattern
- created 3 folders for each CONTROLLER version
