+++ src/main/java/pcd/ass01/Controller/Executor/UpdateBoidPositionTask.java
@@ -1,5 +1,8 @@
-package pcd.ass01;
+package pcd.ass01.Controller.Executor;
 
+
+import pcd.ass01.Model.Boid;
+import pcd.ass01.Model.BoidsModel;
 
 public class UpdateBoidPositionTask extends AbstractBoidTask {
