+++ src/main/java/pcd/ass01/Controller/Executor/UpdateBoidVelocityTask.java
@@ -1,4 +1,7 @@
-package pcd.ass01;
+package pcd.ass01.Controller.Executor;
+
+import pcd.ass01.Model.Boid;
+import pcd.ass01.Model.BoidsModel;
 
 public class UpdateBoidVelocityTask extends AbstractBoidTask {
