+++ src/main/java/pcd/ass01/Model/BoidsModel.java
@@ -1,6 +1,5 @@
-package pcd.ass01;
+package pcd.ass01.Model;
 
-import java.io.Console;
 import java.util.ArrayList;
 import java.util.List;
