+++ src/main/java/pcd/ass01/View/BoidPattern/ShapeDrawer.java
@@ -1,4 +1,4 @@
-package pcd.ass01.BoidPattern;
+package pcd.ass01.View.BoidPattern;
 
 import java.awt.*;
